# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Example Swagger setup for Golang Gin.",
        "title": "Assistant API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9876",
    "basePath": "/",
    "paths": {
...
        "/sys_distributed_lock/acquire": {
            "post": {
                "description": "原子操作：如果锁不存在或已过期则获取成功",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分布式锁"
                ],
                "summary": "尝试获取分布式锁",
                "parameters": [
                    {
                        "description": "锁参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_modules_sys_distributed_lock.acquireReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_distributed_lock/check/{lock_key}": {
            "get": {
                "description": "返回锁是否被持有（未过期）以及holder",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分布式锁"
                ],
                "summary": "检查锁是否被持有",
                "parameters": [
                    {
                        "type": "string",
                        "description": "锁Key",
                        "name": "lock_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_distributed_lock/query/{lock_key}": {
            "get": {
                "description": "获取锁的详细信息（包含holder、TTL等）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分布式锁"
                ],
                "summary": "查询锁信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "锁Key",
                        "name": "lock_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_distributed_lock/release": {
            "post": {
                "description": "释放锁，必须提供 lock_holder 验证所有权",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分布式锁"
                ],
                "summary": "释放分布式锁",
                "parameters": [
                    {
                        "description": "释放参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_modules_sys_distributed_lock.releaseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限（holder不匹配）",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/sys_distributed_lock/renew": {
            "post": {
                "description": "为锁延长TTL，必须验证 lock_holder 是否匹配",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分布式锁"
                ],
                "summary": "续期分布式锁",
                "parameters": [
                    {
                        "description": "续期参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_modules_sys_distributed_lock.renewReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "403": {
                        "description": "无权限（holder不匹配或锁已过期）",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/assistant_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "assistant_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_app_modules_sys_distributed_lock.acquireReq": {
            "type": "object",
            "properties": {
                "lock_holder": {
                    "type": "string"
                },
                "lock_key": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                }
            }
        },
        "internal_app_modules_sys_distributed_lock.releaseReq": {
            "type": "object",
            "properties": {
                "lock_holder": {
                    "type": "string"
                },
                "lock_key": {
                    "type": "string"
                }
            }
        },
        "internal_app_modules_sys_distributed_lock.renewReq": {
            "type": "object",
            "properties": {
                "lock_holder": {
                    "type": "string"
                },
                "lock_key": {
                    "type": "string"
                },
                "ttl": {
                    "type": "integer"
                }
            }
        },
...
    }
}